+++ MyMemo/Utills/NetworkError.swift
@@ -1,0 +1,12 @@
+//
+//  NetworkError.swift
+//  MyMemo
+//
+//  Created by t2023-m0076 on 2023/08/29.
+//
+
+import Foundation
+
+enum NetworkError: Error {
+    case emptyResponse
+}
