+++ MyMemo/ViewController/ViewController.swift
@@ -37,12 +37,7 @@ class ViewController: UIViewController {
                 print("Error: ", error)
             }
         })
-        
+
     }
-     
-    
-    
+
 }
-
-
-
